refactor(tests): move RestartRateLimitMixin to mixins
